Make parallel_tests.sh output failures as they happen.
This is particularly useful if there is a typo or bug in the test
that causes it to fail 100% of the time. In this case, the output
will start with "FFF..." or "EEE..." and the user can interrupt
the script without waiting for the whole run, which may be very
long.

Test: ../net_test/parallel_tests.sh 15 30 privacy_address_test.py
Change-Id: I07b3f7886e5a3d8e5d95ad0b738c6df802101245