Fix Python packages in the bullseye image.
The python package was removed from bullseye. See for example
https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=970420 .

Install the python-is-python3 package to get a /usr/bin/python
symlink back so the shebang lines in the tests continue to work.
We cannot use Python 2 by default because bullseye doesn't have
a Python 2 scapy package, and without scapy the tests are pretty
much useless.

Also install the Python 2 and Python 3 six packages because they
will help transition the tests to Python 3. Currently none of the
tests run, let alone pass, but I have CLs for some of the simpler
ones ready.

Test: ./build_rootfs.sh -a amd64 -s bullseye
Test: ROOTFS=net_test.rootfs.amd64.bullseye.20201112 ../net_test/run_net_test.sh --nobuild cstruct_test.py
Change-Id: I69f8c73735565f7ddc12e9c1a610d1ac54d5a7cf