bpf_test.py: set RLIMIT_MEMLOCK for BpfTest
For some kernels we need to set the RLIMIT_MEMLOCK to large size
for the BPF tests to pass, and according to Maciej Żenczykowski's suggestion,
changing the global limit for a device is not a good idea,
it should be changed just for the test itself.

So here we change RLIMIT_MEMLOCK to 1GB for in the setup for the BpfTest,
and change it back in the tearDown function,
to work aorund the "Operation not permitted" problem which caused by
the RLIMIT_MEMLOCK setting problem.

Test: vts-tradefed run vts-kernel -m vts_kernel_net_tests

Signed-off-by: Yongqin Liu <yongqin.liu@linaro.org>
Change-Id: Ia200fc78f9a115bbc8a578d333e551f3ae59ba4e