Replace qemu-debootstrap with --foreign --second-stage
The qemu-debootstrap method is deprecated now, and it was a trivial
wrapper that can just be expanded in place.

This change starts to pull out some of the filesystem modifications
that are not specific to the OS being installed, but rather the
debootstrap process, so they can be made to work with a multi-pass
QEMU-based solution, instead of relying on qemu-user and chroots.

Change-Id: Ifca7ce9f65329a3aef2e0c0470c89f9282369760