Revert "Add the same mark with policy in SaInfo API"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/43ce5ce4-ead7-4870-b4ef-69e87f4c2ba7). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I09130ce4bbc5853eeb8bc4906fa62063b56570e0