Revert "Add the same mark with policy in SaInfo API"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/5a63e4a6-43fc-4b21-9777-b8da6ff08ab6). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I3b215a50ab91241f20a549d3885e4c24fc233a14