Revert "Add the same mark with policy in SaInfo API"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/2e5cbc0b-8807-4c33-8163-56faa993007b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia9dd6a1a9d3d752a45c57c5b3f05f0bb4a8e2fd1