Revert "net-test: adjust ping6_test.py testProcNetRaw6 for 6.0-rc1"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/7f36f7ce-11e9-4b7c-9581-bad929d0ca40). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ibe0030dea757478c28b58c60ef86c8338b47c2e8