Revert "net-test: adjust ping6_test.py testProcNetRaw6 for 6.0-rc1"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/6c4e85ea-dbbe-4f36-b20f-3729cc6e5c28). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iff415ba90401a4f4e40ff2fa871038a82f42d19c