Revert "net-test: adjust ping6_test.py testProcNetRaw6 for 6.0-rc1"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/cc3b2d3b-4287-4cd9-b933-8bb47bb24638). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I71d456b02442ba5fde7dd5eaa1523f939a38eb8e