Revert "net-test: adjust ping6_test.py testProcNetRaw6 for 6.0-rc1"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/7583c1e6-e10a-4f20-a71e-6340e4f735d7). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7df2b4eb91a1d9e9b476e484db7b3f45928a9d9b