Revert "net-test: adjust ping6_test.py testProcNetRaw6 for 6.0-rc1"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/8d0b65d2-06f7-499e-96f3-4e61cd8b6b50). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia1734995ac34e45d2b56ce180385491b20564f5b