build_rootfs.sh: Enable framebuffer for 'server'
So far we've either worked on headless devices, like rockpi, or the
Cuttlefish device nested, where most console interation happens over the
serial console. But for a real server (like Gigabyte/Ampere) even if we
don't have a monitor connected, the display driver needs to be working
for the built-in BMC/KVM feature to be useful.

When debugging the pre-boot/initramfs environment, it's nice to have
the display driver working so we can use the console there as well.
These drivers would have been loaded by systemd, but for ISO boot we
start with init=/bin/sh, so that never happens. Have the initramfs do
it for us.

Change-Id: I9a77af596ad99389d80ffa33a5d34f352ef84735