Revert "net-test: populate correct procotol in ESP trailer"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/afb4a9d2-0e34-426d-b5d1-6d7f4e584898). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I3f4e0d18ac4a3f471cb19e725fe81dd3b9207ff2