Revert "net-test: adjust 'accept_ra_min_lft' minimum LTS kernel ..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/74948994-45d8-43c0-ba71-2503c01af400). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6f6f75ec989593109cb9ede2c1d885bd3b229b6d