Revert "net-test: kernel_feature_test - temporarily allow 6.1.25"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/036e2196-4747-49cc-a5a0-f1acc5fa2f1c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8b0dbdb43c6ab0ff9be268452591a9c9d6ea1465