Revert "net-test: V VTS - enforce minimum non-buggy LTS kernel v..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/036e2196-4747-49cc-a5a0-f1acc5fa2f1c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib7115fc018b5328d95f8012fe2333154ae0d2776