Update IS_GSI check to use init.gsi.rc file
The /metadata/gsi/dsu/booted file is no longer used to determine if a device is running GSI. Instead, the init.gsi.rc file is used. This file is created by the GSI init script and is only present when GSI is running.

This change will ensure that the net_test script correctly identifies devices that are running GSI.

Change-Id: I6669304726916ef1bac59c7037cb735ee92c0062