Revert "net-test: print out 'hidden' kernel architecture as well"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7efb7552-bd3a-41b6-9f40-3ad43db085c8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iae4ef3ea29b423f9dd8aa6cc711e2825bdcc85f7
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com